Add a scoped search service. Update scope methods
